+++ app/src/main/java/sleepless_nights/location_alarm/MainActivity.java
@@ -1,12 +1,14 @@
-package com.example.locationalarm;
+package sleepless_nights.location_alarm;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.ViewModelProviders;
 
 import android.os.Bundle;
 
-import com.example.locationalarm.alarm.ui.alrmlistfragment.AlarmListFragment;
-import com.example.locationalarm.alarm.view_models.alarm_view_model.AlarmViewModel;
+import com.locationalarm.R;
+
+import sleepless_nights.location_alarm.alarm.ui.alarm_list_fragment.AlarmListFragment;
+import sleepless_nights.location_alarm.alarm.view_models.alarm_view_model.AlarmViewModel;
 
 import java.util.Objects;
