+++ app/src/main/java/sleepless_nights/location_alarm/alarm/Alarm.java
@@ -1,4 +1,4 @@
-package com.example.locationalarm.alarm;
+package sleepless_nights.location_alarm.alarm;
 
 public class Alarm {
     private int id;
